refactor : 전역-타입-분리
