+++ src/entities/user/api/service.ts
@@ -1,5 +1,5 @@
-import type { User } from '../model/types';
 import type { GetUserResponseDTO, GetUsersDTO, GetUsersResponseDTO } from './dto';
+import type { User } from '@/shared/model/types/auth';
 import { apiClient } from '@/shared/api/apiClient';
 import { USER_ENDPOINT } from '@/shared/model/constants/apiEndpoints';
 import { mapUserResponseDTOToUser } from './mappers';
