+++ src/shared/model/store/authStore.ts
@@ -1,6 +1,6 @@
 import { create } from 'zustand';
 import { persist } from 'zustand/middleware';
-import type { LoginStatus, User } from '../types';
+import type { LoginStatus, User } from '../types/auth';
 
 interface AuthState {
   accessToken: string | null;
