+++ src/shared/model/store/chatStore.ts
@@ -1,5 +1,5 @@
 import { create } from 'zustand';
-import type { ChatMessage, SendMessage, UpdateMessage } from '../types';
+import type { ChatMessage, SendMessage, UpdateMessage } from '../types/webSocket';
 
 interface ChatState {
   messages: Record<number, SendMessage[]>;
