+++ src/shared/model/types/webSocket.ts
@@ -1,23 +1,3 @@
-export interface User {
-  id: number;
-  name: string;
-  email: string;
-  profileImageURL: string;
-  loginStatus: LoginStatus;
-  description: string;
-}
-
-export interface Server {
-  serverUri: string;
-  channels: [];
-}
-export interface Channel {
-  channelId: number;
-  name: string;
-}
-
-export type LoginStatus = 'ONLINE' | 'OFFLINE' | 'AWAY' | 'BUSY';
-
 export interface ChatMessage {
   chatId: string;
   recipientId: number;
@@ -56,13 +36,4 @@ export interface WebRTCSignalData {
 export interface WebRTCSignalMessage extends WebSocketMessage {
   type: 'CALL_OFFER' | 'CALL_ANSWER' | 'CALL_ICE' | 'CALL_END';
   data: WebRTCSignalData;
-}
-
-export interface UPdateCHMEssage {
-  content: string;
-}
-
-export interface CHWebSocketMessage {
-  operation: 'SEND' | 'UPDATE' | 'DELETE';
-  type: 'CHANNEL';
 }
